refactor: three core folder structure.
