+++ packages/three/core/index.js
@@ -1,65 +1,11 @@
-import * as THREE from "three";
-import CameraControls from "./helpers/CameraControls";
+export { default as Math } from "./helpers/Math";
+export { raycastGeometry, raycastMeshFromScreen } from "./helpers/Raycaster";
+export { getEmptyRoomGeometry, downloadMesh } from "./RoomGeometry";
 
-function Three({ canvas, alpha = true, interactElement }) {
-  const { width, height } = canvas.getBoundingClientRect();
+export { default as CapturePanorama } from "./CapturePanorama";
+export { create3DRoom } from "./RoomGeometry";
+export { default as TexturePostEffect } from "./TexturePostEffect";
+export { exportTexture } from "./helpers/Texture";
 
-  const scene = new THREE.Scene();
-  const renderer = new THREE.WebGLRenderer({
-    canvas,
-    alpha,
-    preserveDrawingBuffer: true,
-  });
-  renderer.setSize(width, height);
-
-  const camera = new THREE.PerspectiveCamera(75, width / height, 0.01, 1000);
-  const cameraControls = new CameraControls(camera, interactElement);
-
-  const customRender = {};
-  const animate = () => {
-    const animeFrame = requestAnimationFrame(animate);
-    Object.keys(customRender)
-      .map((key) => customRender[key])
-      .forEach((func) => func(renderer));
-    renderer.render(scene, camera);
-
-    return animeFrame;
-  };
-
-  const animeFrame = animate();
-
-  const destroy = () => {
-    cancelAnimationFrame(animeFrame);
-    cameraControls.destroy();
-  };
-
-  const setCanvasSize = (width, height) => {
-    camera.aspect = width / height;
-    camera.updateProjectionMatrix();
-    renderer.setSize(width, height);
-  };
-
-  const addBeforeRenderFunction = (func) => {
-    let id = THREE.MathUtils.generateUUID();
-
-    while (customRender[id] !== undefined) {
-      id = THREE.MathUtils.generateUUID();
-    }
-
-    customRender[id] = func;
-    return () => {
-      delete customRender[id];
-    };
-  };
-
-  return {
-    destroy,
-    setCanvasSize,
-    scene,
-    addBeforeRenderFunction,
-    renderer,
-    cameraControls,
-  };
-}
-
-export default Three;
+export { default as Three } from "./three";
+export { default as Css3D } from "./Css3D";
