+++ packages/three/index.js
@@ -8,20 +8,7 @@ export { default as Css3DObject } from "./components/Css3DObject";
 export { default as TransformControls } from "./components/TransformControls";
 export { TRANSFORM_CONTROLS_MODE } from "./components/TransformControls";
 export { default as MeshIndexMap } from "./components/MeshIndexMap";
+
 export * as Placeholder from "./components/Placeholder";
 export * as Media from "./helpers/MediaLoader";
-
-import Math from "./core/helpers/Math";
-import {
-  raycastGeometry,
-  raycastMeshFromScreen,
-} from "./core/helpers/Raycaster";
-import { getEmptyRoomGeometry, downloadMesh } from "./core/RoomGeometry";
-
-export const Core = {
-  Math,
-  raycastGeometry,
-  raycastMeshFromScreen,
-  getEmptyRoomGeometry,
-  downloadMesh,
-};
+export * as Core from "./core";
